requirements.txt
boto3==1.34.102
botocore==1.34.102
python-dotenv==1.0.0
pandas==1.5.3
Pillow==10.3.0
opencv-python==4.9.0.80
opencv-contrib-python==4.9.0.80
img2table==1.2.10
pytesseract==0.3.10
tenacity==8.2.3
fuzzywuzzy==0.18.0
scikit-image==0.22.0
spacy==3.7.4
presidio-analyzer==2.2.354
presidio-anonymizer==2.2.354
PyYAML==6.0.1
openai==1.44
numpy==1.24.2
tiktoken==0.6.0